refactor(modal): rename the new Modal to Modal
