+++ packages/bezier-react/src/components/Modals/Modal/Modal.styled.ts
