+++ packages/bezier-react/src/components/Modals/Modal/ModalAnimation.styled.ts
